fix: logout
