+++ apps/web/app/future/auth/logout/page.tsx
@@ -1,0 +1,7 @@
+import Logout from "@pages/auth/logout";
+import { withAppDir } from "app/AppDirSSRHOC";
+import { WithLayout } from "app/layoutHOC";
+
+import { getServerSideProps } from "@server/lib/auth/logout/getServerSideProps";
+
+export default WithLayout({ getLayout: null, Page: Logout, getData: withAppDir(getServerSideProps) })<"P">;
